Move UserDaoMapper to service
