+++ src/main/java/service/UserDataAccess.java
@@ -10,7 +10,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
-import user.acceptconnect.UserDaoMapper;
 import user.requestconnect.exceptions.UserNotFoundException;
 
 @Repository
